[fix] 요일 별 운동 계획 리스트에 데이터 통신 구현 중
