+++ lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:projectsampledata/ui/pages/exercise/exercise_plus_page/add_exercise_detail_page.dart';
 import 'package:projectsampledata/ui/pages/main/main_page/main_page.dart';
 import 'package:projectsampledata/ui/pages/splash/splash_page.dart';
 import 'package:projectsampledata/ui/pages/user/user_info_page/user_info_page.dart';
@@ -23,8 +24,7 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
         navigatorKey: navigatorKey,
         debugShowCheckedModeBanner: false,
-        home: SplashPage(),
-        // home: MainPage(),
+        home: MainPage(),
         routes: {
           "/login": (context) => LoginPage(), // 로그인 페이지
           "/join": (context) => JoinPage(),
@@ -34,7 +34,7 @@ class MyApp extends StatelessWidget {
               ListDetailOfDayPage(ClickWeek.dayOfWeekName),
           "/user-info-update-page": (context) => UserInfoUpdatePage(),
         },
-        theme: ThemeData(appBarTheme: appBarTheme()) // 정의한 AppBarTheme 적용 ),
+        theme: ThemeData(appBarTheme: appBarTheme()) // 정의한 AppBarTheme 적용
         );
   }
 }
