+++ lib/ui/pages/week/week_detail_page/widgets/list_detail_of_day_body.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:projectsampledata/data/repository/list_detail_of_day_info_repository.dart';
+import 'package:projectsampledata/data/repository/list_detail_of_day_repository.dart';
 import 'package:projectsampledata/ui/pages/week/week_detail_page/list_detail_of_day_vm.dart';
 import 'package:projectsampledata/ui/pages/week/week_detail_page/widgets/list_detail_of_day_button.dart';
 import 'package:projectsampledata/ui/pages/week/week_detail_page/widgets/list_detail_of_day_card.dart';
