+++ lib/ui/pages/week/week_detail_page/widgets/list_detail_of_day_card.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:projectsampledata/data/repository/list_detail_of_day_info_repository.dart';
+import 'package:projectsampledata/data/repository/list_detail_of_day_repository.dart';
 import 'package:projectsampledata/ui/pages/week/week_detail_page/list_detail_of_day_vm.dart';
 
 class ExerciseCard extends StatelessWidget {
